Rename song/index.ts to song/models.ts
